+++ app/models/shop.rb
@@ -1,6 +1,3 @@
-require "ac_integrator"
-require "slack_notify"
-
 class Shop < ActiveRecord::Base
   has_many :card_orders, dependent: :destroy
   has_many :postcards, through: :card_orders
@@ -30,14 +27,6 @@ class Shop < ActiveRecord::Base
   end
 
   class << self
-
-    # TODO: Should this should live with the session store methods?
-    # TODO: Does this need to be background-tasked?
-    def add_to_email_list(email)
-      ac = AcIntegrator::NewInstall.new
-      ac.add_email_to_list(email)
-    end
-
     # Session store
     def store(session)
       shop = Shop.find_by(domain: session.url)
@@ -47,9 +36,7 @@ class Shop < ActiveRecord::Base
         shop.save!
         shop.get_shopify_id
         shop.sync_shopify_metadata
-        shop.get_last_month
-        add_to_email_list(shop.email)
-        SlackNotify.install(shop.domain, shop.email, shop.owner, shop.last_month, true)
+        AppInstalledJob.perform_later(shop)
       else
         shop.token = session.token
         shop.uninstalled_at = nil
@@ -115,6 +102,9 @@ class Shop < ActiveRecord::Base
     credit - subscriptions.first.quantity
   end
 
+  def increment_credit
+    increment!(:credit)
+  end
 
   # Console admin method for listing all stores that have activated with Stripe
   def self.find_active
@@ -150,9 +140,15 @@ class Shop < ActiveRecord::Base
     }
   end
 
-  def get_last_month
+  def update_last_month
+    # We want to know how many first time customers ordered from the shop. This value can be misleading
+    # if the shop imported data from another platform. We thus cap the number with order/processed_at,
+    # which is the number of orders actually created (not imported) in the last month.
     new_sess
-    last_month = ShopifyAPI::Customer.count(created_at_min: (Time.now - 1.month))
+    start_time = Time.now - 1.month
+    customers_created = ShopifyAPI::Customer.count(created_at_min: start_time)
+    orders_processed = ShopifyAPI::Order.count(processed_at_min: start_time)
+    last_month = [customers_created, orders_processed].min
     update_attribute(:last_month, last_month)
   end
 
@@ -202,6 +198,6 @@ class Shop < ActiveRecord::Base
   
   # necessary for the active admin
   def display_name
-    self.domain
+    self.domain.split('.myshopify.com').first
   end
 end