+++ lib/card_util.rb
@@ -24,7 +24,16 @@ module CardUtil
   # Todo: We should  probably refactor this into `send_card` and share it with `send_card` in postcard.rb
   # we could have another method for sending a promo card (or, better yet, a way to do it from ActiveAdmin)
 
-  def send_promo_card(card_order, lob_to_address, discount_code = 'SAM-PLE-XXX', expiry = 4.weeks.from_now.midnight)
+  def send_promo_card(card_order, lob_to_address)
+
+    discount_code = nil
+    expiry = nil
+
+    if card_order.discount?
+      discount_code = 'SAM-PLE-XXX'
+      expiry = 4.weeks.from_now.midnight
+    end
+
     front_html, back_html = [
         card_order.card_side_front,
         card_order.card_side_back
@@ -33,8 +42,8 @@ module CardUtil
           background_image: card_side.image,
           discount_x: card_side.discount_x,
           discount_y: card_side.discount_y,
-          discount_pct: card_order.discount_pct,
-          discount_exp: expiry.strftime("%m/%d/%Y"),
+          discount_pct: card_order.discount_pct ? card_order.discount_pct.abs : nil,
+          discount_exp: expiry ? expiry.strftime("%m/%d/%Y") : nil,
           discount_code: card_side.show_discount? ? discount_code : nil
       )
     end
@@ -45,7 +54,9 @@ module CardUtil
         to: lob_to_address,
         # from: "Sent by Touchcard\nhttp://Touchcard.co",  # Does not pass address verification
         front: front_html,
-        back: back_html
+        back: back_html,
+        # Allow time for canceling if sample is messed up
+        send_date: 1.hour.from_now.iso8601
     )
   end
 end